requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.11.7
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
asyncpg==0.28.0
python-dotenv==1.0.0
email-validator==2.1.0.post1
sentence-transformers==2.2.2
numpy>=2.3.1
pytest==8.4.1
httpx==0.25.1
pytest-asyncio==0.21.1
black==23.10.1
isort==5.12.0
flake8==6.1.0
greenlet==3.1.1

# AI and Vector Database Dependencies
langchain>=0.1.0
langchain_openai==0.3.28
langchain_pinecone==0.2.9
langchain_core>=0.3.70
pinecone==7.3.0

# Speech-to-Text Dependencies
openai-whisper==20231117

